requirements.txt
discord.py==1.7.3
#tweepy==3.10.0
requests==2.26.0
toml==0.10.2